+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.0.1
 
-- Initial release of the internationalization package.
+- Initial release of the lang_internationalization package.
 - Automatically generate localization files for multiple languages.
 - Merge new keys with existing localization files without overwriting existing keys.
 - Easy to use annotations for specifying supported locales.